feat: fortalece operacao e resiliencia financeira
Adiciona conciliacao bidirecional, tratamento auditavel de quarentena, contratos versionados, protecao contra resultados ambiguos, robustez de notificacoes e novos cenarios de seguranca e recuperacao.

Atualiza a documentacao com as evidencias de 202 testes, DAST, carga, consistencia distribuida e operacao local.

+++ services/notification-service/src/main/java/br/com/orquestrapay/notification/data/RepositorioWebhooks.java
@@ -268,7 +268,7 @@ public class RepositorioWebhooks {
                         """)
                 .param("idEmpresa", idEmpresa)
                 .param("limite", tamanho)
-                .param("deslocamento", pagina * tamanho)
+                .param("deslocamento", (long) pagina * tamanho)
                 .query(this::mapearResposta)
                 .list();
         return new PaginaEntregasWebhook(itens, pagina, tamanho, total);
